bidirectional Clips fully working
