+++ patches/girlpower/Ableton/Control (Message Ableton).v4p
@@ -1,5 +1,5 @@
 <!DOCTYPE PATCH  SYSTEM "http://vvvv.org/versions/vvvv45beta34.2.dtd" >
-   <PATCH nodename="C:\dev\vvvv\vvvv-Message\patches\girlpower\Ableton\Control (Message Ableton).v4p" systemname="Control (Message Ableton)" filename="C:\dev\vvvv\vvvv-Message\patches\girlpower\Ableton\Control (Message Ableton).v4p" scrollx="0" scrolly="2625">
+   <PATCH nodename="C:\dev\vvvv\vvvv-Message\patches\girlpower\Ableton\Control (Message Ableton).v4p" systemname="Control (Message Ableton)" filename="C:\dev\vvvv\vvvv-Message\patches\girlpower\Ableton\Control (Message Ableton).v4p" scrollx="1020" scrolly="2670">
    <NODE nodename="IOBox (Value Advanced)" componentmode="InABox" id="0" systemname="IOBox (Value Advanced)">
    <BOUNDS type="Box" left="4395" top="1500" width="480" height="480">
    </BOUNDS>
@@ -427,9 +427,9 @@
    <PACK Name="addonpack" Version="34.1.0">
    </PACK>
    <NODE nodename="IOBox (Value Advanced)" componentmode="InABox" id="20" systemname="IOBox (Value Advanced)">
-   <BOUNDS type="Box" left="12015" top="8895" width="285" height="240">
+   <BOUNDS type="Box" left="14490" top="8850" width="285" height="240">
    </BOUNDS>
-   <BOUNDS type="Node" left="12015" top="8895" width="0" height="0">
+   <BOUNDS type="Node" left="14490" top="8850" width="0" height="0">
    </BOUNDS>
    <PIN pinname="Y Input Value" slicecount="1" visible="1" values="1">
    </PIN>
@@ -455,11 +455,13 @@
    </PIN>
    <PIN pinname="Precision" slicecount="1" values="4">
    </PIN>
+   <PIN pinname="Descriptive Name" slicecount="1" values="time">
+   </PIN>
    </NODE>
    <NODE systemname="IOBox (Value Advanced)" nodename="IOBox (Value Advanced)" componentmode="InABox" id="25">
-   <BOUNDS type="Node" left="10935" top="8370" width="100" height="100">
+   <BOUNDS type="Node" left="11040" top="8085" width="100" height="100">
    </BOUNDS>
-   <BOUNDS type="Box" left="10935" top="8370" width="480" height="480">
+   <BOUNDS type="Box" left="11040" top="8085" width="480" height="480">
    </BOUNDS>
    <PIN pinname="Value Type" slicecount="1" values="Boolean">
    </PIN>
@@ -470,6 +472,8 @@
    <PIN pinname="Y Input Value" slicecount="1" values="0">
    </PIN>
    <PIN pinname="Y Output Value" visible="1">
+   </PIN>
+   <PIN pinname="Descriptive Name" slicecount="1" values="overdub">
    </PIN>
    </NODE>
    <NODE nodename="IOBox (Value Advanced)" componentmode="InABox" id="26" systemname="IOBox (Value Advanced)">
@@ -485,9 +489,11 @@
    </PIN>
    <PIN pinname="Y Output Value" visible="1">
    </PIN>
+   <PIN pinname="Descriptive Name" slicecount="1" values="tempo">
+   </PIN>
    </NODE>
    <NODE systemname="Edit (Message Formular)" filename="%VVVV%\packs\vvvv-Message\nodes\plugins\VVVV.Nodes.Messaging.dll" nodename="Edit (Message Formular)" componentmode="InAWindow" id="19">
-   <BOUNDS type="Node" left="8340" top="9255" width="100" height="100">
+   <BOUNDS type="Node" left="8340" top="9255" width="6600" height="270">
    </BOUNDS>
    <PIN pinname="Message Formular" slicecount="1" visible="1" values="Ableton.Live">
    </PIN>
@@ -499,17 +505,17 @@
    </PIN>
    <PIN pinname="AutoSense" visible="1" pintype="Input" slicecount="1" values="0">
    </PIN>
-   <PIN pinname="master.pan" slicecount="1" values="0.15">
+   <PIN pinname="master.pan" slicecount="1" values="0.15" visible="1">
    </PIN>
-   <PIN pinname="track" slicecount="1" values="-5">
+   <PIN pinname="track" slicecount="1" values="-5" visible="1">
    </PIN>
    <PIN pinname="overdub" slicecount="1" visible="1" values="2">
    </PIN>
-   <PIN pinname="master.volume" slicecount="1" values="0.63">
+   <PIN pinname="master.volume" slicecount="1" values="0.63" visible="1">
    </PIN>
    <PIN pinname="time" visible="1" slicecount="1" values="1.76">
    </PIN>
-   <PIN pinname="master.crossfader" slicecount="1" values="0.06">
+   <PIN pinname="master.crossfader" slicecount="1" values="0.06" visible="1">
    </PIN>
    <PIN pinname="tempo" visible="1" slicecount="1" values="247">
    </PIN>
@@ -532,4 +538,74 @@
    </NODE>
    <LINK srcnodeid="28" srcpinname="Output Enum" dstnodeid="19" dstpinname="Message Formular">
    </LINK>
+   <NODE nodename="IOBox (Value Advanced)" componentmode="InABox" id="29" systemname="IOBox (Value Advanced)">
+   <BOUNDS type="Box" left="12270" top="8055" width="795" height="240">
+   </BOUNDS>
+   <BOUNDS type="Node" left="12270" top="8055" width="0" height="0">
+   </BOUNDS>
+   <PIN pinname="Y Input Value" slicecount="1" visible="1" values="-1">
+   </PIN>
+   <PIN pinname="Units" slicecount="1" values="||">
+   </PIN>
+   <PIN pinname="Precision" slicecount="1" values="4">
+   </PIN>
+   <PIN pinname="Value Type" slicecount="1" values="Integer">
+   </PIN>
+   <PIN pinname="Descriptive Name" slicecount="1" values="track">
+   </PIN>
+   </NODE>
+   <LINK srcnodeid="29" srcpinname="Y Output Value" dstnodeid="19" dstpinname="track">
+   </LINK>
+   <NODE nodename="IOBox (Value Advanced)" componentmode="InABox" id="30" systemname="IOBox (Value Advanced)">
+   <BOUNDS type="Box" left="13200" top="8310" width="795" height="240">
+   </BOUNDS>
+   <BOUNDS type="Node" left="13200" top="8310" width="0" height="0">
+   </BOUNDS>
+   <PIN pinname="Y Input Value" slicecount="1" visible="1" values="0.06">
+   </PIN>
+   <PIN pinname="Units" slicecount="1" values="||">
+   </PIN>
+   <PIN pinname="Precision" slicecount="1" values="4">
+   </PIN>
+   <PIN pinname="Descriptive Name" slicecount="1" values="master.crossfader">
+   </PIN>
+   </NODE>
+   <LINK srcnodeid="30" srcpinname="Y Output Value" dstnodeid="19" dstpinname="master.crossfader">
+   </LINK>
+   <NODE nodename="IOBox (Value Advanced)" componentmode="InABox" id="31" systemname="IOBox (Value Advanced)">
+   <BOUNDS type="Box" left="16035" top="7785" width="795" height="240">
+   </BOUNDS>
+   <BOUNDS type="Node" left="16035" top="7785" width="0" height="0">
+   </BOUNDS>
+   <PIN pinname="Y Input Value" slicecount="1" visible="1" values="0.3">
+   </PIN>
+   <PIN pinname="Units" slicecount="1" values="||">
+   </PIN>
+   <PIN pinname="Precision" slicecount="1" values="4">
+   </PIN>
+   <PIN pinname="Descriptive Name" slicecount="1" values="master.volume">
+   </PIN>
+   </NODE>
+   <LINK srcnodeid="31" srcpinname="Y Output Value" dstnodeid="19" dstpinname="master.volume">
+   </LINK>
+   <NODE nodename="IOBox (Value Advanced)" componentmode="InABox" id="32" systemname="IOBox (Value Advanced)">
+   <BOUNDS type="Box" left="15240" top="7095" width="795" height="240">
+   </BOUNDS>
+   <BOUNDS type="Node" left="15240" top="7095" width="0" height="0">
+   </BOUNDS>
+   <PIN pinname="Y Input Value" slicecount="1" visible="1" values="0.15">
+   </PIN>
+   <PIN pinname="Units" slicecount="1" values="||">
+   </PIN>
+   <PIN pinname="Precision" slicecount="1" values="4">
+   </PIN>
+   <PIN pinname="Descriptive Name" slicecount="1" values="master.pan">
+   </PIN>
+   </NODE>
+   <LINK srcnodeid="32" srcpinname="Y Output Value" dstnodeid="19" dstpinname="master.pan">
+   </LINK>
+   <LINK srcnodeid="10" srcpinname="Output" dstnodeid="19" dstpinname="Input">
+   </LINK>
+   <PACK Name="vvvv-ZeroMQ" Version="0.5.0">
+   </PACK>
    </PATCH>
