+++ src/MessageNodes/VVVV.Packs.Messaging.Nodes.csproj
@@ -124,12 +124,15 @@
     <Compile Include="Utils\MessageCloneNode.cs" />
     <Compile Include="Spread\MessageMergeNode.cs" />
     <Compile Include="Utils\MessageChangeNode.cs" />
+    <Compile Include="Utils\Fields\MessageContainsFieldNode.cs" />
     <Compile Include="Utils\MessageFrameDelayNode.cs" />
     <Compile Include="Keep\MessageKeepHoldNode.cs">
       <SubType>UserControl</SubType>
     </Compile>
     <Compile Include="Utils\MessageInfoNode.cs" />
-    <Compile Include="Utils\MessagePruneNode.cs" />
+    <Compile Include="Utils\Fields\MessagePruneEmptyNode.cs" />
+    <Compile Include="Utils\Fields\MessagePruneIndicatedNode.cs" />
+    <Compile Include="Utils\Fields\MessagePruneNode.cs" />
     <Compile Include="Spread\MessageSiftNode.cs" />
     <Compile Include="Keep\Abstract\TypableMessageKeepNode.cs">
       <SubType>UserControl</SubType>
@@ -154,9 +157,9 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Spread\MessageSortNode.cs" />
     <Compile Include="Utils\MessageStreamingNodes.cs" />
-    <Compile Include="Utils\MessageTopicNode.cs" />
+    <Compile Include="Utils\Fields\MessageTopicNode.cs" />
     <Compile Include="Utils\MessageRemoveNode.cs" />
-    <Compile Include="Utils\MessageRenameField.cs" />
+    <Compile Include="Utils\Fields\MessageRenameField.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
