+++ src/OSC/Enum.cs
@@ -1,0 +1,18 @@
+﻿namespace VVVV.Nodes.OSC
+{
+    public enum SelectEnum
+    {
+        All,
+        First,
+        Last
+    }
+
+    enum Filter {
+        Matches,
+        Contains,
+        Starts,
+        Ends,
+        All
+    }
+    
+}
